cmd: minor unit and flag improvements
- `get-transaction` on miner tx's now reveal the outputs in `m`
- `get-block` now includes a `--last` to allow one to easily say "get me
the last block", or "get me the last-1 block" ...
- `get-last-block-header` now shows rewards in precise `m`

+++ pkg/rpc/daemon/types.go
@@ -774,7 +774,7 @@ type TransactionJSON struct {
 		} `json:"key"`
 	} `json:"vin"`
 	Vout []struct {
-		Amount int `json:"amount"`
+		Amount uint64 `json:"amount"`
 		Target struct {
 			Key string `json:"key"`
 		} `json:"target"`
